safari-534.53-branch/LayoutTests/fast/forms/textarea-selection-preservation-expected.txt

This tests for problems where we'd lose the selection in a textarea when making style and value changes.

PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- add background style
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- set value to same value
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- set value to a different value
PASS ta.selectionStart is 6
PASS ta.selectionEnd is 6
- set selection so we can test again without a trailing newline
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- add background style
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- set value to same value
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
PASS successfullyParsed is true

TEST COMPLETE
